+++ chariBE/src/main/java/com/macia/chariBE/model/DonateDetails.java
@@ -26,7 +26,7 @@ public class DonateDetails {
     private Integer DND_ID;
 
     @Column
-    private Double money;
+    private int money;
 
     @Column
     private String message;
